fix: error is thrown when no `.less` files are found
In case there are no `.less` files in the application, the plugin's hook will fail with error `The "path" argument must be one of type string, Buffer, or URL. Received type undefined`
The problem is that we iterrate over elements of empty array, so the `arr[0]` is undefined and the `path.join` operation fails.
Fix this by skipping the transpilation when there aren't `.less` files in the application.