fix: support projects with nsconfig.json
Support projects with nsconfig.json files - currently they cannot be used as the `nativescript-dev-less` plugin always checks `app` dir for `.less` files.
Also fix the check if the files is from `App_Resources` dir - currently it is not working on Windows as the `glob` always returns UNIX style paths, but we compare them with Windows style ones.